Refactor feval and subsref.
Also, change the shape of output when evaluating
row-valued chebfuns at arrays of points so it is
made up of vertically stacked blocks of the same
shape as the point array, rather than vertically
stacked blocks each in the shape of the point array
with the first two dimensions transposed.

cf. https://github.com/chebfun/chebfun/issues/1606